Introduce the mandatory test suite  that a supported logging library MUST satisfy for log injection
